refactor :: character controller command와 query 분리, 이후 service를 비롯한 다른 계층도 분리하였으면 좋겠음
